+++ CHANGELOG.md
@@ -45,6 +45,7 @@
     - support recursive glob (a.k.a ``globstar``)
         - enabled via ``globstar`` runtime option (enabled by default)
     - now propagate ``opendir`` error, such as ``EMFILE``, ``ENFILE``, ``ENOMEM``
+    - check glob recursion depth
 - **Breaking Change**: now not preserve exit status during finally/defer block
     - now the following code is valid
       ```